[Ref] MP3 Samples: Silence -Wconversion warnings. [Ref] Vorbis Samples: Silence -Wconversion warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@18565 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/SampleFormatMP3.cpp
@@ -598,7 +598,7 @@ bool CSoundFile::ReadMP3Sample(SAMPLEINDEX sample, FileReader &file, bool raw, b
 	{
 		m_szNames[sample] = mpt::ToCharset(GetCharsetInternal(), sampleName);
 		Samples[sample].Initialize();
-		Samples[sample].nC5Speed = rate;
+		Samples[sample].nC5Speed = static_cast<uint32>(rate);
 	}
 	Samples[sample].nLength = mpt::saturate_cast<SmpLength>((data.size() / channels) - data_skip_frames);
